Fix issuedAt type assertion

+++ util/session/sessionmanager.go
@@ -129,7 +129,7 @@ func (mgr *SessionManager) Parse(tokenString string) (jwt.Claims, error) {
 		return nil, err
 	}
 
-	issuedAt := time.Unix(claims["iat"].(int64), 0)
+	issuedAt := time.Unix(int64(claims["iat"].(float64)), 0)
 	if issuedAt.Before(mgr.settings.AdminPasswordMtime) {
 		return nil, fmt.Errorf("Password for superuser has changed since token issued")
 	}
